+++ vtex-my-subscriptions-3/react/typings/graphql/queries/search.gql.d.ts
@@ -2,19 +2,16 @@ declare module '*/search.gql' {
   import { DocumentNode } from 'graphql'
 
   interface SearchItem {
+    name: string
     skuId: string
     imageUrl: string
-    variations: string[]
     plans: string[]
   }
   export interface SearchProduct {
-    productId: string
-    productName: string
     unitMultiplier: number
     measurementUnit: string
     brand: string
     price: number
-    categories: string[]
     items: SearchItem[]
   }